// src/pages/workflowConfig/workflow/nodes/knowledge-retrieval/types.ts
import type { CommonNodeType, ModelConfig, ValueSelector } from '@/pages/workflowConfig/workflow/types'
import type { RETRIEVE_TYPE } from '@/pages/workflowConfig/types/app'
import type {
  DataSet,
  RerankingModeEnum,
} from '@/pages/workflowConfig/models/datasets'

export type MultipleRetrievalConfig = {
  top_k: number
  score_threshold: number | null | undefined
  reranking_model?: {
    provider: string
    model: string
  }
  reranking_mode?: RerankingModeEnum
  weights?: {
    vector_setting: {
      vector_weight: number
      embedding_provider_name: string
      embedding_model_name: string
    }
    keyword_setting: {
      keyword_weight: number
    }
  }
  reranking_enable?: boolean
}

export type SingleRetrievalConfig = {
  model: ModelConfig
}

export type KnowledgeRetrievalNodeType = CommonNodeType & {
  query_variable_selector: ValueSelector
  dataset_ids: string[]
  retrieval_mode: RETRIEVE_TYPE
  multiple_retrieval_config?: MultipleRetrievalConfig
  single_retrieval_config?: SingleRetrievalConfig
  retrieval_model?: any
  _datasets?: DataSet[]
}